adding mimetype and reusing filename image content properties
