+++ matrix/services/message/src/main/kotlin/app/dapk/st/matrix/message/internal/SendMessageUseCase.kt
@@ -38,21 +38,15 @@ internal class SendMessageUseCase(
             }
 
             is MessageService.Message.ImageMessage -> {
-                println("Sending message")
                 val imageContent = imageContentReader.read(message.content.uri)
-
-                println("content: ${imageContent.size}")
-
-                val uri = httpClient.execute(uploadRequest(imageContent.content, imageContent.fileName, "image/png")).contentUri
-                println("Got uri $uri")
-
+                val uri = httpClient.execute(uploadRequest(imageContent.content, imageContent.fileName, imageContent.mimeType)).contentUri
                 val request = sendRequest(
                     roomId = message.roomId,
                     eventType = EventType.ROOM_MESSAGE,
                     txId = message.localId,
                     content = MessageService.Message.Content.ImageContent(
                         url = uri,
-                        filename = "foobar.png",
+                        filename = imageContent.fileName,
                         MessageService.Message.Content.ImageContent.Info(
                             height = imageContent.height,
                             width = imageContent.width,
@@ -65,39 +59,4 @@ internal class SendMessageUseCase(
         }
     }
 
-}
-
-
-interface ImageContentReader {
-    fun read(uri: String): ImageContent
-
-    data class ImageContent(
-        val height: Int,
-        val width: Int,
-        val size: Long,
-        val fileName: String,
-        val content: ByteArray
-    ) {
-        override fun equals(other: Any?): Boolean {
-            if (this === other) return true
-            if (javaClass != other?.javaClass) return false
-
-            other as ImageContent
-
-            if (height != other.height) return false
-            if (width != other.width) return false
-            if (size != other.size) return false
-            if (!content.contentEquals(other.content)) return false
-
-            return true
-        }
-
-        override fun hashCode(): Int {
-            var result = height
-            result = 31 * result + width
-            result = 31 * result + size.hashCode()
-            result = 31 * result + content.contentHashCode()
-            return result
-        }
-    }
 }
