+++ matrix/services/sync/src/main/kotlin/app/dapk/st/matrix/sync/internal/sync/RoomEventFactory.kt
@@ -44,7 +44,6 @@ internal class RoomEventFactory(
 
     private fun ApiTimelineEvent.TimelineMessage.readImageMeta(userCredentials: UserCredentials): RoomEvent.Image.ImageMeta {
         val content = this.content as ApiTimelineEvent.TimelineMessage.Content.Image
-        println(content)
         return RoomEvent.Image.ImageMeta(
             content.info?.width,
             content.info?.height,
